migrate defaultProps
